Refactored controllers and added throwing exceptions to repositories.

+++ src/repository/UserRepository.php
@@ -5,15 +5,16 @@ require_once __DIR__ . '/../models/User.php';
 
 class UserRepository extends Repository
 {
-    public function getUser(string $email): ?User
+
+    public function getUser(string $email): User
     {
         $statement = $this->database->connect()->prepare('
         SELECT * FROM public.users_details_view WHERE email = :email');
-        $statement->bindParam(':email', $email, PDO::PARAM_STR);
+        $statement->bindParam(':email', $email);
         $statement->execute();
         $user = $statement->fetch(PDO::FETCH_ASSOC);
         if ($user == false) {
-            return null;
+            throw new UnexpectedValueException("User with this email doesn't exist");
         }
         $userObj = new User($user['email'], $user['login'], $user['password']);
         $userObj->setName($user['name']);
@@ -27,20 +28,20 @@ class UserRepository extends Repository
         $role = Utility::USER;
         $statement = $this->database->connect()->prepare('
         SELECT * FROM public.users_details_view WHERE role = :role ORDER BY login');
-        $statement->bindParam(':role', $role, PDO::PARAM_STR);
+        $statement->bindParam(':role', $role);
         $statement->execute();
         return $statement->fetchAll(PDO::FETCH_ASSOC);
     }
 
-    public function getIdByEmail(string $email): ?array
+    public function getIdByEmail(string $email): array
     {
         $statement = $this->database->connect()->prepare('
         SELECT id FROM public.users WHERE email = :email');
-        $statement->bindParam(':email', $email, PDO::PARAM_STR);
+        $statement->bindParam(':email', $email);
         $statement->execute();
         $id = $statement->fetch(PDO::FETCH_ASSOC);
         if ($id == false) {
-            return null;
+            throw new UnexpectedValueException('id not found');
         }
         return $id;
     }
@@ -56,7 +57,7 @@ class UserRepository extends Repository
         $statement3 = $this->database->connect()->prepare('
         SELECT id FROM public.users_details WHERE login = :login
         ');
-        $statement3->bindParam(':login', $login, PDO::PARAM_STR);
+        $statement3->bindParam(':login', $login);
         $statement3->execute();
         $foundId = $statement3->fetch(PDO::FETCH_ASSOC);
 
@@ -73,6 +74,9 @@ class UserRepository extends Repository
         $statement->bindParam(':id', $id, PDO::PARAM_INT);
         $statement->execute();
         $foundId = $statement->fetch(PDO::FETCH_ASSOC);
+        if ($foundId == false) {
+            throw new UnexpectedValueException('User not found');
+        }
         $statement2 = $this->database->connect()->prepare('
         DELETE FROM public.users_details WHERE id = :id');
         $statement2->bindParam(':id', $foundId['users_details_id'], PDO::PARAM_INT);
@@ -86,8 +90,8 @@ class UserRepository extends Repository
         $statement = $this->database->connect()->prepare('
         SELECT * FROM public.users_details_view WHERE lower(login) like :string AND  role like :role ORDER BY login
         ');
-        $statement->bindParam(':string', $searchString, PDO::PARAM_STR);
-        $statement->bindParam(':role', $role, PDO::PARAM_STR);
+        $statement->bindParam(':string', $searchString);
+        $statement->bindParam(':role', $role);
         $statement->execute();
         return $statement->fetchAll(PDO::FETCH_ASSOC);
 
@@ -97,7 +101,7 @@ class UserRepository extends Repository
     {
         $statement = $this->database->connect()->prepare('
         SELECT email FROM public.users_details_view WHERE email = :email');
-        $statement->bindParam(':email', $email, PDO::PARAM_STR);
+        $statement->bindParam(':email', $email);
         $statement->execute();
         $found = $statement->fetch(PDO::FETCH_ASSOC);
         if (!$found) {
@@ -110,7 +114,7 @@ class UserRepository extends Repository
     {
         $statement = $this->database->connect()->prepare('
         SELECT login FROM public.users_details_view WHERE login = :login');
-        $statement->bindParam(':login', $login, PDO::PARAM_STR);
+        $statement->bindParam(':login', $login);
         $statement->execute();
         $found = $statement->fetch(PDO::FETCH_ASSOC);
         if (!$found) {
@@ -126,7 +130,7 @@ class UserRepository extends Repository
         UPDATE public.users SET  password = :password WHERE id = :id
         ');
             $passwordHash = password_hash($password, PASSWORD_DEFAULT);
-            $statement->bindParam(':password', $passwordHash, PDO::PARAM_STR);
+            $statement->bindParam(':password', $passwordHash);
             $statement->bindParam(':id', $id, PDO::PARAM_INT);
             $statement->execute();
         }
@@ -134,15 +138,15 @@ class UserRepository extends Repository
         $statement3 = $this->database->connect()->prepare('
         SELECT users_details_id FROM public.users WHERE email = :email
         ');
-        $statement3->bindParam(':email', $email, PDO::PARAM_STR);
+        $statement3->bindParam(':email', $email);
         $statement3->execute();
         $foundId = $statement3->fetch(PDO::FETCH_ASSOC);
 
         $statement2 = $this->database->connect()->prepare('
         UPDATE public.users_details SET login = :login, name = :name WHERE  id = :id
         ');
-        $statement2->bindParam(':login', $login, PDO::PARAM_STR);
-        $statement2->bindParam(':name', $name, PDO::PARAM_STR);
+        $statement2->bindParam(':login', $login);
+        $statement2->bindParam(':name', $name);
         $statement2->bindParam(':id', $foundId['users_details_id'], PDO::PARAM_INT);
         $statement2->execute();
     }
